+++ Frontend/src/components/Atoms/Typography/H2/H2.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './H2.scss';
 
-const StyledH2 = ({color, children}) => {
+const StyledH2 = ({color="", children}) => {
     return <h2 className={`styledH2 ${color}`}>{children}</h2>
 }