+++ Frontend/src/views/NewSetView/NewSetView.js
@@ -1,12 +1,12 @@
-import React, {useRef} from 'react';
+import React from 'react';
 import MainTemplate from '../../templates/MainTemplate';
 import './NewSetView.scss';
 import StyledH2 from '../../components/Atoms/Typography/H2/H2';
 import StyledInput from '../../components/Atoms/StyledInput/StyledInput';
 import Button from '../../components/Atoms/Button/Button';
+import NewFisheItem from '../../components/Organisms/NewFicheItem/NewFisheItem';
 
 const NewSetView = () => {
-    const titleRef = useRef(null);
 
     const createNewSet = () => {
         console.log('Work');
@@ -16,17 +16,21 @@ const NewSetView = () => {
     <MainTemplate>
         <div className="newSetWrapper">
             <StyledH2 color="black">Create new set :</StyledH2>
+            <div className="titleAndDescriptionWrapper">
             <StyledInput
                 id="Title"
-                ref={titleRef}
                 placeholder={`Enter title, for example : "Javascript is the Best".`}
                 labelValue="Title" />
             <StyledInput
                 id="setDesription"
-                ref={titleRef}
-                placeholder={`Additional description, "Words to test"`}
-                labelValue="Title" />
-            <Button onClickFn={createNewSet}>Create new Set</Button>
+                placeholder={`Add additional description, "Vocabulary to French test."`}
+                labelValue="Description" />
+                
+            </div>
+            <div className="rightTopPosition">
+               <Button onClickFn={createNewSet}>Create</Button>
+            </div>
+            <NewFisheItem> </NewFisheItem>
         </div>
     </MainTemplate>
     );