+++ changelog.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2020.06.12
+
+**Added**
+- Svgs: Trash.svg, drag.svg,
+- Components: ButtonIcon, newFisheItem.
+
+## [0.0.4] - 2020.06.11
+
+**Changes**
+- Expaned StyledInput for easier development.
+
+**Added**
+- Components : Button.
+
 ## [0.0.3] - 2020.06.10
 
 **Added**